Country based AQI comparision tool
It is intend to ease the understadning extent of difference between AQI
followed in different countries. This app trying to show how the AQI
values are differeing based on the pollutant concentration range,
variablity of AQI category and colour code.

It is purely written in html and javascript following the AQI calculator
provided by EPA, US.

It comprised of 1) tick boxes for choosing the variable, 2) text box to
enter the values3) buttons to reset and calculate, 3) a table to give
output for AQI value, colour code and AQI remark for different countries.
The variables choosen are PM2.5, PM10, NO2, SO2 mostly to cover the
NAMP data of India. Number of countries to be compared for variable
PM10 is 14, for PM2.5 is 7, SO2 is 9 and NO2 is 5. With more info on AQI comparision need, an option to compare AQI
with historical data from NAMP data.

First step going to be done is edit the EPA page to compare single
variable for all countries. Step2 for all variables. Step3 adding
the styling using the bootstrap. Step3 add the D3 calendar to compare
the NAMP data AQI with AQI output based on any one another country's
AQI output.